+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Expose static with on `UInt`, `Int` & `U8aFixed` classes with optional type name override
 - Support for arbitrary UInt types via `UInt<bitLength>` definitions
 - Add `api.injectMetadata(metadata)` to initialize with a specific version
+- The api now checks for runtime upgrades, decorating with new metadata
 
 # 1.3.1 Feb 18, 2020